adding money to smartcontract
we need to add money to our smart bank. we will use the add money function
as this function has a special keyword payable. it means we are depositing money ot our contract.

msg.sender is the that person who will add the money into our contract with msg.value which is amount of ether